change ReportData : ReportData = {company:{}, currentUser:{}, startpage:.. } removekey password out of user data
change retaileruser doc/collection. it uses _user db
add more routes on adminRouter.js

+++ terminal-installer/attachments/reporting/admins/adminRouter.js
@@ -65,14 +65,19 @@ var adminRouter =
 	     {routes: {
 		  "menuAdministration/companyReport":"menuAdministrationCompany",
 		  "menuAdministration/groupReport":"menuAdministrationGroup",
-		  "menuAdministration/storeReport":"menuAdministrationStore"
+		  "menuAdministration/storeReport":"menuAdministrationStore",
+		  "companyReport/group/:group_id/admins" :"menuAdministrationCompany",
+		  "companyReport/store/:store_id/admins" :"menuAdministrationCompany",
+		  "groupReport/store/:store_id/admins" :"menuAdministrationGroup"
 	      },
-	      menuAdministrationCompany:function() {
-		  this._setup("companyReport",ReportData.company._id);
+	      menuAdministrationCompany:function(id) {
+          var levelid = _.isEmpty(id)?ReportData.company._id:id;
+		  this._setup("companyReport",levelid);
 		  console.log("menuAdministrationCompany");
 	      },
-	      menuAdministrationGroup:function() {
-		  this._setup("groupReport",ReportData.group.group_id);
+	      menuAdministrationGroup:function(id) {
+          var levelid = _.isEmpty(id)?ReportData.group.group_id:id;
+		  this._setup("groupReport",levelid);
 		  console.log("menuAdministrationGroup");
 	      },
 	      menuAdministrationStore:function() {
@@ -80,7 +85,7 @@ var adminRouter =
 		  console.log("menuAdministrationStore");
 	      },
 	      _setup:function(startPage, id){
-		  fetchRetailerUserCollection_Report(id, ReportData.currentUser._id)(function(err,collection){
+		  fetchRetailerUserCollection_Report(id, ReportData)(function(err,collection){
 						      var view = this.view;
 						      var html = ich.adminManagement_TMP(_.extend({startPage:startPage},autoBreadCrumb()));
 						      $("#main").html(html);
@@ -105,11 +110,11 @@ var menuAdminUsersView =
 				 
 				 $("#usersInfoTable").html(ich.adminUsersInfotable_TMP({list:list}));
 				 _.each(list,function(item){
-					    $("#edit-"+item._id).button().click(function(){
+					    $("#edit-"+item.name).button().click(function(){
 										    //alert("edit");
 										    var userModel = collection.get(item._id);
 										    var userJSON = userModel.toJSON();
-										    var data_TMP = {ismaster:(userJSON.role=="MASTERADMIN"?true:false)
+										    var data_TMP = {ismaster:isBackOfficeAdminUser(userJSON)
 										                   ,isactive:(userJSON.status=="ACTIVE"?true:false)};
                                             _.extend(data_TMP,userJSON);
 										    
@@ -120,10 +125,11 @@ var menuAdminUsersView =
                                                                                  ,userJSON:userJSON}));
 										});
 					    
-					    $("#del-"+item._id).button().click(function(){
+					    $("#del-"+item.name).button().click(function(){
 										   //alert("del");
 										   var adminModel = collection.get(item._id);
-										   if(_.isNotEmpty(adminModel) && (adminModel.toJSON())._id!=ReportData.currentUser._id) {
+										   if(_.isNotEmpty(adminModel) && (adminModel.toJSON())._id!=ReportData.currentUser._id
+										                               && (adminModel.toJSON())._id!=(_.last(list))._id) {
 										       adminModel.destroy({
 										           success:function() {
 										               view._renderTable(collection);
@@ -134,36 +140,36 @@ var menuAdminUsersView =
 										           }
 										       });
 										   } else {
-										       alert("Can't delete Current User");
+										       alert("Can't delete This User");
 										   }
 									       });
 					});
 				 $("#addusers")
 				     .button()
 				     .click(function(){
-						function getDefaultData(ReportData) {
-						    if(ReportData.company) {
-							return {company_id:ReportData.company._id,
-								companyName:ReportData.company.companyName};
-						    } else if(ReportData.group) {
-							return {company_id:ReportData.company_id,
-								companyName:ReportData.companyName,
-								group_id:ReportData.group.group_id,
-								groupName:ReportData.group.groupName};
-						    } else if(ReportData.store) {
-							return {company_id:ReportData.company_id,
-								companyName:ReportData.companyName,
-								group_id:ReportData.group_id,
-								groupName:ReportData.groupName,
-								store_id:ReportData.store.store_id,
-								storeName:ReportData.store.storeName,
-								storeNumber:ReportData.store.number};
+						function getDefaultData(reportData) {
+						    if(reportData.company) {
+							return {company_id:reportData.company._id,
+								companyName:reportData.company.companyName};
+						    } else if(reportData.group) {
+							return {company_id:reportData.company_id,
+								companyName:reportData.companyName,
+								group_id:reportData.group.group_id,
+								groupName:reportData.group.groupName};
+						    } else if(reportData.store) {
+							return {company_id:reportData.company_id,
+								companyName:reportData.companyName,
+								group_id:reportData.group_id,
+								groupName:reportData.groupName,
+								store_id:reportData.store.store_id,
+								storeName:reportData.store.storeName,
+								storeNumber:reportData.store.number};
 						    } else {
 							return undefined;
 						    }
 						};
 						
-						if(ReportData.currentUser.role=="MASTERADMIN") {
+						if(isBackOfficeAdminUser(ReportData.currentUser)) {
     						quickCreateUserInfoDialog(ich.inputUserInfo_TMP({})
     									   ,_.extend(addUser(getDefaultData(ReportData),collection)
     									           ,{collection:collection}));
